# apps/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Open Blog API
  description: |
    A comprehensive REST API for a blog platform built with NestJS and Prisma.

    ## Features
    - User authentication with JWT tokens
    - Blog post management (CRUD operations)
    - Categories and tags support
    - Initial setup workflow
    - Admin dashboard capabilities

    ## Authentication
    Most admin endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: Open Blog API Support
    email: support@openblog.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate a user and receive a JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            example:
              username: 'admin'
              password: 'password123'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Get current user
      description: Get the current authenticated user's information
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posts/slug/{slug}:
    get:
      summary: Get post by slug
      description: Retrieve a specific post by its slug (public endpoint)
      tags:
        - Posts (Public)
      parameters:
        - name: slug
          in: path
          required: true
          description: The post slug
          schema:
            type: string
            example: 'my-first-blog-post'
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/posts:
    get:
      summary: Get all posts (admin)
      description: Retrieve all posts including drafts (admin only)
      tags:
        - Posts (Admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlogPost'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new post
      description: Create a new blog post (admin only)
      tags:
        - Posts (Admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/posts/{id}:
    get:
      summary: Get post by ID
      description: Retrieve a specific post by its ID (admin only)
      tags:
        - Posts (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The post ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update a post
      description: Update an existing blog post (admin only)
      tags:
        - Posts (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The post ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a post
      description: Delete a blog post (admin only)
      tags:
        - Posts (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The post ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Post deleted successfully'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories:
    get:
      summary: Get all categories
      description: Retrieve all available categories (public endpoint)
      tags:
        - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      summary: Create a new category
      description: Create a new category (admin only)
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /categories/{id}:
    get:
      summary: Get category by ID
      description: Retrieve a specific category by its ID
      tags:
        - Categories
      parameters:
        - name: id
          in: path
          required: true
          description: The category ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update a category
      description: Update an existing category (admin only)
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The category ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a category
      description: Delete a category (admin only)
      tags:
        - Categories
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The category ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Category deleted successfully'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags:
    get:
      summary: Get all tags
      description: Retrieve all available tags
      tags:
        - Tags
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new tag
      description: Create a new tag (Admin only)
      tags:
        - Tags
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagDto'
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Tag name or slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags/{id}:
    get:
      summary: Get tag by ID
      description: Retrieve a specific tag by its ID
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      responses:
        '200':
          description: Tag found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update tag
      description: Update an existing tag (Admin only)
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagDto'
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Tag name or slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete tag
      description: Delete an existing tag (Admin only)
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tag ID
      responses:
        '204':
          description: Tag deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete tag - it is being used by posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tags/slug/{slug}:
    get:
      summary: Get tag by slug
      description: Retrieve a specific tag by its slug
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Tag slug
      responses:
        '200':
          description: Tag found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username or email
          example: 'admin'
        password:
          type: string
          description: User password
          example: 'password123'
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
        user:
          $ref: '#/components/schemas/User'
...
    CreatePostDto:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: Post title
          example: 'My First Blog Post'
        content:
          type: string
          description: Post content (HTML/Markdown)
          example: '<p>This is the content of my first blog post.</p>'
        excerpt:
          type: string
          description: Short excerpt of the post
          example: 'A brief summary of the post'
        featuredImage:
          type: string
          description: URL to the featured image
          example: 'https://example.com/image.jpg'
        published:
          type: boolean
          description: Whether the post is published
          default: false
          example: true
        publishedAt:
          type: string
          format: date-time
          description: When the post was published
          example: '2025-07-30T15:30:00Z'
        metaTitle:
          type: string
          description: SEO meta title
          example: 'My First Blog Post - Open Blog'
        metaDescription:
          type: string
          description: SEO meta description
          example: 'Learn about creating your first blog post'
        categoryId:
          type: string
          format: uuid
          description: Category ID
          example: '123e4567-e89b-12d3-a456-426614174000'
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          description: Array of tag IDs
          example:
            [
              '123e4567-e89b-12d3-a456-426614174001',
              '123e4567-e89b-12d3-a456-426614174002',
            ]
    UpdatePostDto:
      type: object
      properties:
        title:
          type: string
          description: Post title
          example: 'Updated Blog Post Title'
        content:
          type: string
          description: Post content (HTML/Markdown)
          example: '<p>Updated content of the blog post.</p>'
        excerpt:
          type: string
          description: Short excerpt of the post
          example: 'Updated brief summary'
        featuredImage:
          type: string
          description: URL to the featured image
          example: 'https://example.com/updated-image.jpg'
        published:
          type: boolean
          description: Whether the post is published
          example: true
        publishedAt:
          type: string
          format: date-time
          description: When the post was published
          example: '2025-07-30T16:00:00Z'
        metaTitle:
          type: string
          description: SEO meta title
          example: 'Updated Blog Post - Open Blog'
        metaDescription:
          type: string
          description: SEO meta description
          example: 'Updated description for SEO'
        categoryId:
          type: string
          format: uuid
          description: Category ID
          example: '123e4567-e89b-12d3-a456-426614174000'
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          description: Array of tag IDs
          example: ['123e4567-e89b-12d3-a456-426614174001']
...
    CreateCategoryDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Category name
          example: 'Technology'
        description:
          type: string
          maxLength: 500
          description: Category description
          example: 'Posts about technology and programming'
        color:
          type: string
          maxLength: 7
          description: Category color (hex code)
          example: '#3B82F6'
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Category name
          example: 'Updated Technology'
        description:
          type: string
          maxLength: 500
          description: Category description
          example: 'Updated posts about technology and programming'
        color:
          type: string
          maxLength: 7
          description: Category color (hex code)
          example: '#10B981'
    BlogPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique post identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        title:
          type: string
          description: Post title
          example: 'My First Blog Post'
        slug:
          type: string
          description: URL-friendly post slug
          example: 'my-first-blog-post'
        content:
          type: string
          description: Post content
          example: '<p>This is the content of my blog post.</p>'
        excerpt:
          type: string
          nullable: true
          description: Short excerpt
          example: 'A brief summary'
        featuredImage:
          type: string
          nullable: true
          description: Featured image URL
          example: 'https://example.com/image.jpg'
        published:
          type: boolean
          description: Publication status
          example: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
          description: Publication date
          example: '2025-07-30T15:30:00Z'
        createdAt:
          type: string
          format: date-time
          description: Creation date
          example: '2025-07-30T15:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update date
          example: '2025-07-30T15:30:00Z'
        metaTitle:
          type: string
          nullable: true
          description: SEO meta title
          example: 'My First Blog Post - Open Blog'
        metaDescription:
          type: string
          nullable: true
          description: SEO meta description
          example: 'Learn about creating your first blog post'
        authorId:
          type: string
          format: uuid
          description: Author user ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        categoryId:
          type: string
          format: uuid
          nullable: true
          description: Category ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        author:
          $ref: '#/components/schemas/User'
        category:
          allOf:
            - $ref: '#/components/schemas/Category'
            - nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: '123e4567-e89b-12d3-a456-426614174001'
        email:
          type: string
          format: email
          description: User email
          example: 'admin@example.com'
        username:
          type: string
          description: Username
          example: 'admin'
        firstName:
          type: string
          nullable: true
          description: First name
          example: 'John'
        lastName:
          type: string
          nullable: true
          description: Last name
          example: 'Doe'
        role:
          type: string
          enum: [ADMIN, AUTHOR]
          description: User role
          example: 'ADMIN'
        isActive:
          type: boolean
          description: Whether user is active
          example: true
        isTemporaryPassword:
          type: boolean
          description: Whether user has a temporary password that must be changed
          example: false
        mustChangePassword:
          type: boolean
          description: Whether user must change password on next login
          example: false
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp
          example: '2025-08-05T15:30:00Z'
        passwordResetAt:
          type: string
          format: date-time
          nullable: true
          description: Last password reset timestamp
          example: '2025-08-05T14:00:00Z'
        passwordChangedAt:
          type: string
          format: date-time
          nullable: true
          description: Last password change timestamp
          example: '2025-08-05T14:05:00Z'
        createdAt:
          type: string
          format: date-time
          description: Account creation date
          example: '2025-07-30T15:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update date
          example: '2025-07-30T15:00:00Z'
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique category identifier
          example: '123e4567-e89b-12d3-a456-426614174002'
        name:
          type: string
          description: Category name
          example: 'Technology'
        slug:
          type: string
          description: URL-friendly category slug
          example: 'technology'
        description:
          type: string
          nullable: true
          description: Category description
          example: 'Posts about technology and programming'
        color:
          type: string
          nullable: true
          description: Category color (hex code)
          example: '#3B82F6'
        createdAt:
          type: string
          format: date-time
          description: Creation date
          example: '2025-07-30T15:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update date
          example: '2025-07-30T15:00:00Z'
...
    CreateTagDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Tag name (slug will be auto-generated)
          minLength: 1
          maxLength: 50
          example: 'JavaScript'
    UpdateTagDto:
      type: object
      properties:
        name:
          type: string
          description: Updated tag name
          minLength: 1
          maxLength: 50
          example: 'JavaScript Framework'
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Error message(s)
          example: 'Invalid input data'
        error:
          type: string
          description: Error type
          example: 'Bad Request'

    # User Management Schemas
...